requirements.txt
absl-py==1.0.0
alabaster==0.7.12
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blessed==1.20.0
blis==0.2.4
boto3==1.24.2
botocore==1.27.2
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
conllu==1.3.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docutils==0.19
editdistance==0.6.0
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.8.0
flaky==3.7.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.32.0
ftfy==6.1.1
fuzzywuzzy==0.18.0
gensim==3.8.3
gevent==22.10.1
greenlet==1.1.3.post0
grpcio==1.43.0
h5py==3.7.0
huggingface-hub==0.10.1
idna==3.4
imagesize==1.4.1
importlib-metadata==4.10.0
importlib-resources==5.10.0
iniconfig==1.1.1
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.2.0
jsonnet==0.18.0
jsonpickle==2.2.0
jsonschema==4.17.0
kiwisolver==1.4.4
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==2.0.4
murmurhash==1.0.7
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.4.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
numpydoc==1.5.0
nvidia-ml-py==12.535.108
oauthlib==3.1.1
overrides==3.1.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parsimonious==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
plac==0.9.6
pluggy==1.0.0
preshed==2.0.1
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.1
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pygtrie==2.4.2
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.1.3
python-dateutil==2.8.2
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2022.5
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.2.1
rapidfuzz==2.13.2
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.0
responses==0.21.0
rouge==1.0.1
rsa==4.8
s3transfer==0.6.0
sacremoses==0.0.46
scikit-learn==1.1.2
scipy==1.9.2
Send2Trash==1.8.0
sentencepiece==0.1.96
setuptools==59.5.0
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
srsly==1.0.5
stack-data==0.5.1
synonyms==3.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.5
termcolor==1.1.0
terminado==0.17.0
thinc==7.0.8
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.23.1
typing_extensions==4.2.0
typing-utils==0.1.0
Unidecode==1.3.4
urllib3==1.26.12
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.3
word2number==1.1
xdg==5.1.1
zipp==3.9.0
zope.event==4.5.0
zope.interface==5.5.0